support maps. These are from JsonConverter with `schemas.enable=false`
