+++ DEVELOPMENT.md
@@ -33,11 +33,24 @@ Follow instructions in [Releasing](https://central.sonatype.org/publish/release/
 This was tested on a Confluent quickstart distribution downloaded from Confluent
 website.
 
+The following test scenarios use the confluent `FileStream` connectors. They are not enabled by default in the
+confluent distributions.
+
+To enable them:
+
+1. Find your CONFLUENT_HOME directory. This is where your confluent distribution is unzipped/untarred.
+2. Next, copy ${CONFLUENT_HOME}/share/filestream-connectors ${CONFLUENT_HOME}/share/java/filestream-connectors
+3. Restart the connect server
+4. Make sure that the connectors are installed by running `curl -s http://localhost:8083/connector-plugins` and
+verifying that `FileSource` and `FileSink` connectors show up
+
 # Test With Sink Connector
 
-## Connector Configuration
+This was tested using a simple FileSink connector. The configurations depend on the converter implementation used
 
-This was tested using a simple FileSink connector:
+## StringConverter 
+
+### Config
 
 ```json
 {
@@ -50,26 +63,78 @@ This was tested using a simple FileSink connector:
     "topics": "json-topic",
     "transforms": "mask_json_field",
     "transforms.mask_json_field.type": "io.github.ferozed.kafka.connect.transforms.MaskJsonField$Value",
-    "transforms.mask_json_field.OUTER_FIELD_PATH": "",
-    "transforms.mask_json_field.MASK_FIELD_NAME": "ssn",
+    "transforms.mask_json_field.REPLACEMENT_FIELD_PATH": "/ssn",
 
-    "errors.tolerance": "all"
+    "errors.tolerance": "all",
+    "errors.log.enable": "true",
+    "errors.log.include.messages": "true"
 
 }
 ```
 
-## Deploy connector
+### Deploy connector
 
 ```bash
-curl -X PUT -H "content-type: application/json" http://localhost:8083/connectors/json-mask-test/config -d @/home/ferozed/file-sink.json
+curl -X PUT \
+-H "content-type: application/json" \
+http://localhost:8083/connectors/json-mask-test/config \
+-d @/home/ferozed/file-sink.json
 ```
 
-## Produce JSON Records
+### Produce JSON Records
 
 ```bash
-./confluent-7.0.0/bin/kafka-console-producer --bootstrap-server localhost:9092 --topic json-topic
+./confluent-7.0.0/bin/kafka-console-producer \
+--bootstrap-server \
+localhost:9092 \
+--topic json-topic
 
 { "name": "jon", "ssn": "111-22-3333" }
+```
+
+## JsonConverter
+
+### Config
+```json
+{
+    "name": "json-mask-test-json-converter",
+    "connector.class": "FileStreamSink",
+    "file": "/tmp/file-sink.txt",
+    "key.converter": "org.apache.kafka.connect.storage.StringConverter",
+    "value.converter": "org.apache.kafka.connect.json.JsonConverter",
+    "value.converter.schemas.enable": "false",
+    "tasks.max": 1,
+    "topics": "json-topic",
+    "transforms": "mask_json_field",
+    "transforms.mask_json_field.type": "io.github.ferozed.kafka.connect.transforms.MaskJsonField$Value",
+    "transforms.mask_json_field.REPLACEMENT_FIELD_PATH": "/ssn",
+    "transforms.mask_json_field.CONNECT_FIELD_NAME": "document",
+
+    "errors.tolerance": "all",
+    "errors.log.enable": "true",
+    "errors.log.include.messages": "true"
+
+}
+```
+
+### Deploy connector
+
+```bash
+curl -X PUT \
+-H "content-type: application/json" \
+http://localhost:8083/connectors/json-mask-test/config \
+-d @/home/ferozed/file-sink.json
+```
+
+### Produce JSON Records
+
+```bash
+./confluent-7.0.0/bin/kafka-console-producer \
+--bootstrap-server \
+localhost:9092 \
+--topic json-topic
+
+{"document": "{ \"name\": \"jon\", \"ssn\": \"111-22-3333\" }"}
 ```
 
 ## Validate the output
@@ -78,9 +143,11 @@ Look at the `/tmp/file-sink.txt` file and make sure that the output is correct.
 
 # Test With Source Connector
 
-## Connector Configuration
+## StringConverter
 
-This was tested using a simple FileStreamSource connector:
+### Connector Configuration
+
+This was tested using a simple FileStreamSource connector.
 
 ```json
 {
@@ -93,31 +160,42 @@ This was tested using a simple FileStreamSource connector:
   "topic": "json-source-topic",
   "transforms": "mask_json_field",
   "transforms.mask_json_field.type": "io.github.ferozed.kafka.connect.transforms.MaskJsonField$Value",
-  "transforms.mask_json_field.OUTER_FIELD_PATH": "",
-  "transforms.mask_json_field.MASK_FIELD_NAME": "ssn",
+  "transforms.mask_json_field.REPLACEMENT_FIELD_PATH": "/ssn",
 
   "errors.tolerance": "all",
-  "errors.log.enable": "true"
+  "errors.log.enable": "true",
+  "errors.log.include.messages": "true"
 
 }
 ```
 
-## Deploy connector
+### Deploy connector
 
 ```bash
-curl -X PUT -H "content-type: application/json" http://localhost:8083/connectors/json-mask-test-source/config -d @/home/ferozed/file-source.json
+curl -X PUT \
+-H "content-type: application/json" \
+http://localhost:8083/connectors/json-mask-test-source/config \
+-d @/home/ferozed/file-source.json
 ```
 
-## Produce JSON Records
+### Produce JSON Records
 
 ```bash
-for i in {1..1000}; do echo "{\"name\": \"jon-${i}\", \"ssn\": \"111-22-${i}\"}" >> /tmp/json-input.txt;done
+for i in {1..1000}; \
+do \
+echo "{\"name\": \"jon-${i}\", \"ssn\": \"111-22-${i}\"}" >> /tmp/json-input.txt;\
+done
 ```
 
-## Consume from source topic
+### Consume from source topic
 
 ```bash
-./confluent-7.0.0/bin/kafka-console-consumer --bootstrap-server localhost:9092 --topic json-source-topic --partition 0 --offset 0```
+./confluent-7.0.0/bin/kafka-console-consumer \
+--bootstrap-server \
+localhost:9092 \
+--topic json-source-topic \
+--partition 0 \
+--offset 0
 ```
 
 ## Validate the output
