+++ build.gradle
@@ -31,7 +31,7 @@ repositories {
 }
 
 group = 'io.github.ferozed.kafka.connect'
-version = '1.0'
+version = '1.1'
 
 dependencies {
