+++ src/main/java/io/github/ferozed/kafka/connect/transforms/MaskJsonFieldConfig.java
@@ -1,5 +1,5 @@
 /**
- * Copyright © Feroze Daud (ferozed DOT oss AT gmail.com)
+ * Copyright © 2023 Feroze Daud (ferozed DOT oss AT gmail.com)
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
