+++ src/test/java/io/github/ferozed/kafka/connect/transforms/MaskJsonFieldTest.java
@@ -20,12 +20,6 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.*;
-import com.google.common.collect.ImmutableMap;
-import org.apache.kafka.connect.connector.ConnectRecord;
-import org.apache.kafka.connect.data.Schema;
-import org.apache.kafka.connect.data.SchemaBuilder;
-import org.apache.kafka.connect.data.Struct;
-import org.apache.kafka.connect.sink.SinkRecord;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
@@ -34,325 +28,7 @@ import java.math.BigInteger;
 import java.util.HashMap;
 import java.util.Map;
 
-public class MaskJsonFieldTest {
-
-    ObjectMapper mapper = new ObjectMapper();
-
-    @Test
-    public void testSimpleStringInValue() throws JsonProcessingException {
-        MaskJsonField maskJsonField = new MaskJsonField.Value();
-        maskJsonField.configure(
-                ImmutableMap.of(
-                        MaskJsonFieldConfig.REPLACEMENT_FIELD_PATH, "/foo/bar"
-                )
-        );
-
-        Schema valueSchema = SchemaBuilder.STRING_SCHEMA;
-
-
-        ObjectNode node = (ObjectNode)mapper.createObjectNode();
-
-        node.set("foo",
-                mapper.createObjectNode().put("bar", "ssn")
-                );
-
-        String value = mapper.writeValueAsString(node);
-
-        SinkRecord sinkRecord = new SinkRecord(
-                "topic",
-                0,
-                SchemaBuilder.STRING_SCHEMA,
-                "key",
-                SchemaBuilder.STRING_SCHEMA,
-                value,
-                0
-        );
-
-        ConnectRecord transformedRecord = maskJsonField.apply(sinkRecord);
-
-        assertValue((String)transformedRecord.value(), "/foo/bar", "");
-    }
-
-    @Test
-    public void testSimpleStringInKey() throws JsonProcessingException {
-        MaskJsonField maskJsonField = new MaskJsonField.Key();
-        maskJsonField.configure(
-                ImmutableMap.of(
-                        MaskJsonFieldConfig.REPLACEMENT_FIELD_PATH, "/foo/bar"
-                )
-        );
-
-        Schema valueSchema = SchemaBuilder.STRING_SCHEMA;
-        Schema keySchema = SchemaBuilder.STRING_SCHEMA;
-
-
-        ObjectNode node = (ObjectNode)mapper.createObjectNode();
-
-        node.set("foo",
-                mapper.createObjectNode().put("bar", "ssn")
-        );
-
-        String value = mapper.writeValueAsString(node);
-
-        SinkRecord sinkRecord = new SinkRecord(
-                "topic",
-                0,
-                SchemaBuilder.STRING_SCHEMA,
-                value,
-                SchemaBuilder.STRING_SCHEMA,
-                value,
-                0
-        );
-
-        ConnectRecord transformedRecord = maskJsonField.apply(sinkRecord);
-
-        assertValue((String)transformedRecord.value(), "/foo/bar", "ssn");
-        assertValue((String)transformedRecord.key(), "/foo/bar", "");
-    }
-
-    @Test
-    public void testSimpleStrinWithoutReplacement() throws JsonProcessingException {
-        MaskJsonField maskJsonField = new MaskJsonField.Value();
-        maskJsonField.configure(
-                ImmutableMap.of(
-                        MaskJsonFieldConfig.REPLACEMENT_FIELD_PATH, "/foo/bar"
-                )
-        );
-
-        Schema valueSchema = SchemaBuilder.STRING_SCHEMA;
-
-
-        ObjectNode node = (ObjectNode)mapper.createObjectNode();
-
-        String value = "{\"foo\":{\"nothing\":\"1\"}}";
-
-        SinkRecord sinkRecord = new SinkRecord(
-                "topic",
-                0,
-                SchemaBuilder.STRING_SCHEMA,
-                "key",
-                SchemaBuilder.STRING_SCHEMA,
-                value,
-                0
-        );
-
-        ConnectRecord transformedRecord = maskJsonField.apply(sinkRecord);
-
-        Assertions.assertEquals(value, (String)transformedRecord.value());
-    }
-
-    @Test
-    public void testInvalidJsonWithoutReplacement() throws JsonProcessingException {
-        MaskJsonField maskJsonField = new MaskJsonField.Value();
-        maskJsonField.configure(
-                ImmutableMap.of(
-                        MaskJsonFieldConfig.REPLACEMENT_FIELD_PATH, "/foo/bar"
-                )
-        );
-
-        Schema valueSchema = SchemaBuilder.STRING_SCHEMA;
-
-
-        ObjectNode node = (ObjectNode)mapper.createObjectNode();
-
-        String value = "{";
-
-        SinkRecord sinkRecord = new SinkRecord(
-                "topic",
-                0,
-                SchemaBuilder.STRING_SCHEMA,
-                "key",
-                SchemaBuilder.STRING_SCHEMA,
-                value,
-                0
-        );
-
-        ConnectRecord transformedRecord = maskJsonField.apply(sinkRecord);
-
-        Assertions.assertEquals(value, (String)transformedRecord.value());
-    }
-
-    @Test
-    public void testEmptyJson() throws JsonProcessingException {
-        MaskJsonField maskJsonField = new MaskJsonField.Value();
-        maskJsonField.configure(
-                ImmutableMap.of(
-                        MaskJsonFieldConfig.REPLACEMENT_FIELD_PATH, "/foo/bar"
-                )
-        );
-
-        Schema valueSchema = SchemaBuilder.STRING_SCHEMA;
-
-
-        ObjectNode node = (ObjectNode)mapper.createObjectNode();
-
-        String value = "{}";
-
-        SinkRecord sinkRecord = new SinkRecord(
-                "topic",
-                0,
-                SchemaBuilder.STRING_SCHEMA,
-                "key",
-                SchemaBuilder.STRING_SCHEMA,
-                value,
-                0
-        );
-
-        ConnectRecord transformedRecord = maskJsonField.apply(sinkRecord);
-
-        Assertions.assertEquals(value, (String)transformedRecord.value());
-    }
-
-    @Test
-    public void testToplevelFieldInConnectKeyStruct() throws JsonProcessingException {
-        testToplevelFieldInConnectStructLib(true, "111-22-3333", "");
-    }
-
-    @Test
-    public void testToplevelFieldInConnectValueStruct() throws JsonProcessingException {
-        testToplevelFieldInConnectStructLib(false, "111-22-3333", "");
-    }
-
-    @Test
-    public void testToplevelFieldInConnectValueStructInt() throws JsonProcessingException {
-        testToplevelFieldInConnectStructLib(false, 1234, 0);
-    }
-
-    @Test
-    public void testToplevelFieldInConnectValueStructBigIntMax() throws IOException {
-        testLib(
-                "{\"name\":\"john\",\"ssn\":" + Long.valueOf(Long.MAX_VALUE).toString() + "}",
-                "/ssn",
-                -1L,
-                "{\"name\":\"john\",\"ssn\":-1}"
-        );
-    }
-
-    @Test
-    public void testToplevelFieldInConnectValueStructFloat() throws JsonProcessingException {
-        testToplevelFieldInConnectStructLib(false, 1234.0, 0.0);
-    }
-
-    @Test
-    private void testToplevelFieldInConnectStructLib(boolean isKey, Object valueToReplace, Object expectedReplacement) throws JsonProcessingException {
-        MaskJsonField maskJsonField = isKey ? new MaskJsonField.Key() : new MaskJsonField.Value();
-        maskJsonField.configure(
-                ImmutableMap.of(
-                        MaskJsonFieldConfig.CONNECT_FIELD_NAME, "document",
-                        MaskJsonFieldConfig.REPLACEMENT_FIELD_PATH, "/ssn"
-                )
-        );
-
-        Schema documentSchema = SchemaBuilder.struct()
-                .field("document", Schema.STRING_SCHEMA)
-                .schema();
-
-        Schema valueSchema = SchemaBuilder.struct()
-                .field("database", Schema.STRING_SCHEMA)
-                .field("document", Schema.STRING_SCHEMA
-                ).schema();
-
-        String jsonPayloadTemplate = "{\"first_name\": \"john\", \"last_name\": \"doe\", \"ssn\": VALUE_TO_REPLACE}";
-        String jsonDocument = "";
-        if (valueToReplace.getClass() == String.class) {
-            jsonDocument = jsonPayloadTemplate.replace("VALUE_TO_REPLACE", "\"" + ((String) valueToReplace) + "\"");
-        } else {
-            jsonDocument = jsonPayloadTemplate.replace("VALUE_TO_REPLACE", valueToReplace.toString());
-        }
-        Struct value = new Struct(valueSchema)
-                .put("database", "dynamo")
-                .put("document", jsonDocument);
-
-        SinkRecord sinkRecord = null;
-
-        if (isKey) {
-            sinkRecord = new SinkRecord(
-                    "topic",
-                    0,
-                    valueSchema,
-                    value,
-                    SchemaBuilder.STRING_SCHEMA,
-                    "value",
-                    0
-            );
-        } else {
-            sinkRecord = new SinkRecord(
-                    "topic",
-                    0,
-                    SchemaBuilder.STRING_SCHEMA,
-                    "key",
-                    valueSchema,
-                    value,
-                    0
-            );
-        }
-
-        ConnectRecord transformedRecord = maskJsonField.apply(sinkRecord);
-
-        Struct data = isKey ? (Struct)transformedRecord.key(): (Struct)transformedRecord.value();
-        String documentJson = data.getString("document");
-        assertValue(documentJson, "/ssn", expectedReplacement);
-    }
-
-    @Test
-    public void testNestedFieldInConnectStruct() throws JsonProcessingException {
-        MaskJsonField maskJsonField = new MaskJsonField.Value();
-        maskJsonField.configure(
-                ImmutableMap.of(
-                        MaskJsonFieldConfig.CONNECT_FIELD_NAME, "inner.document",
-                        MaskJsonFieldConfig.REPLACEMENT_FIELD_PATH, "/ssn"
-                )
-        );
-
-        Schema innerSchema = SchemaBuilder.struct()
-                .field("document", Schema.STRING_SCHEMA)
-                .schema();
-
-        Schema valueSchema = SchemaBuilder.struct()
-                .field("inner", innerSchema)
-                .field("database", Schema.STRING_SCHEMA
-                ).schema();
-
-        Struct value = new Struct(valueSchema)
-                .put("database", "dynamo")
-                .put("inner",
-                        new Struct(innerSchema)
-                                .put("document", "{\"first_name\": \"john\", \"last_name\": \"doe\", \"ssn\": \"122-12-1212\"}")
-                );
-
-        SinkRecord sinkRecord = new SinkRecord(
-                "topic",
-                0,
-                SchemaBuilder.STRING_SCHEMA,
-                "key",
-                valueSchema,
-                value,
-                0
-        );
-
-        ConnectRecord transformedRecord = maskJsonField.apply(sinkRecord);
-
-        Struct data = (Struct)transformedRecord.value();
-        String documentJson = data.getStruct("inner").getString("document");
-        assertValue(documentJson, "/ssn", "");
-    }
-
-    void assertValue(String jsonPayload, String path, Object value) throws JsonProcessingException {
-        JsonNode root = mapper.readTree(jsonPayload);
-        JsonPointer pointer = JsonPointer.compile(path);
-        JsonNode targetNode = root.at(pointer);
-
-        if (value.getClass() == String.class) {
-            Assertions.assertTrue(targetNode.isTextual() && targetNode.textValue().equals(value), String.format("%s in %s = \"%s\" Actual=\"%s\"", path, jsonPayload, value, targetNode.textValue()));
-        } else if (value.getClass() == Integer.class) {
-            Assertions.assertTrue(Integer.valueOf(targetNode.intValue()).equals(value), String.format("%s in %s = \"%s\" Actual=\"%s\"", path, jsonPayload, value, targetNode.textValue()));
-        } else if (value.getClass() == Double.class) {
-            Assertions.assertTrue( Double.valueOf(targetNode.doubleValue()).equals(value), String.format("%s in %s = \"%s\" Actual=\"%s\"", path, jsonPayload, value, targetNode.textValue()));
-        } else if (value.getClass() == Float.class) {
-            Assertions.assertTrue(Float.valueOf(targetNode.floatValue()).equals(value), String.format("%s in %s = \"%s\" Actual=\"%s\"", path, jsonPayload, value, targetNode.textValue()));
-        }
-    }
-
+public class MaskJsonFieldTest extends BaseTests {
     @Test
     void jsonPointerTest() throws Exception {
         String payload = "{\"foo\": { \"bar\": \"ssn\" }}";
@@ -432,130 +108,6 @@ public class MaskJsonFieldTest {
         System.out.println(mapper.writeValueAsString(root));
     }
 
-    @Test
-    public void arrayTests() throws IOException {
-        testLib(
-                "{\"ssn\": \"111\"}",
-                "/ssn",
-                "",
-                "{\"ssn\":\"\"}"
-        );
-
-        testLib(
-                "{\"ssn\": \"111\"}",
-                "/ssn",
-                "_REDACTED_",
-                "{\"ssn\":\"_REDACTED_\"}"
-        );
-
-        // REPLACE WHOLE ARRAY
-        testLib(
-                "{\"ssn\": [\"111\",\"22\",\"3333\" ]}",
-                "/ssn",
-                "_REDACTED_",
-                "{\"ssn\":[]}"
-        );
-
-        testLib(
-                "{\"ssn\": [\"111\",\"22\",\"3333\" ]}",
-                "/ssn/2",
-                "_REDACTED_",
-                "{\"ssn\":[\"111\",\"22\",\"_REDACTED_\"]}"
-        );
-
-        // array in deep struct
-        testLib(
-                "{\"foo\":{\"bar\":{\"ssn\":[\"111\",\"22\",\"3333\"]}}}",
-                "/foo/bar/ssn/2",
-                "_REDACTED_",
-                "{\"foo\":{\"bar\":{\"ssn\":[\"111\",\"22\",\"_REDACTED_\"]}}}"
-        );
-
-        // array of int in deep struct
-        testLib(
-                "{\"foo\":{\"bar\":{\"ssn\":[111,22,3333]}}}",
-                "/foo/bar/ssn/2",
-                -1,
-                "{\"foo\":{\"bar\":{\"ssn\":[111,22,-1]}}}"
-        );
-
-        // array of float in deep struct
-        testLib(
-                "{\"oo\":{\"bar\":{\"ssn\":[111.1,22.2,3333.2]}}}",
-                "/oo/bar/ssn/2",
-                -1.1,
-                "{\"oo\":{\"bar\":{\"ssn\":[111.1,22.2,-1.1]}}}"
-        );
-
-    }
-
-    private void testLib(
-            String payload,
-            String path,
-            Object replacement,
-            String expectedJson
-    ) throws IOException {
-        MaskJsonField maskJsonField = new MaskJsonField.Value();
-
-        Map<String,Object> configs = new HashMap<>();
-
-        configs.put(MaskJsonFieldConfig.CONNECT_FIELD_NAME, "inner.document");
-        configs.put(MaskJsonFieldConfig.REPLACEMENT_FIELD_PATH, path);
-
-        if (replacement.getClass() == Integer.class) {
-            configs.put(MaskJsonFieldConfig.REPLACEMENT_VALUE_INT, (Integer)replacement);
-        }
-        else if (replacement.getClass() == Long.class) {
-            configs.put(MaskJsonFieldConfig.REPLACEMENT_VALUE_LONG, (Long)replacement);
-        }
-        else if (replacement.getClass() == Float.class) {
-            configs.put(MaskJsonFieldConfig.REPLACEMENT_VALUE_DOUBLE, Double.valueOf((Float)replacement));
-        }
-        else if (replacement.getClass() == Double.class) {
-            configs.put(MaskJsonFieldConfig.REPLACEMENT_VALUE_DOUBLE, (Double)replacement);
-        }
-        else if (replacement.getClass() == String.class) {
-            configs.put(MaskJsonFieldConfig.REPLACEMENT_VALUE_STRING, (String)replacement);
-        } else {
-            throw new RuntimeException("Unsupported value type: " + replacement.getClass());
-        }
-
-        maskJsonField.configure(configs);
-
-        Schema innerSchema = SchemaBuilder.struct()
-                .field("document", Schema.STRING_SCHEMA)
-                .schema();
-
-        Schema valueSchema = SchemaBuilder.struct()
-                .field("inner", innerSchema)
-                .field("database", Schema.STRING_SCHEMA
-                ).schema();
-
-        Struct value = new Struct(valueSchema)
-                .put("database", "dynamo")
-                .put("inner",
-                        new Struct(innerSchema)
-                                .put("document", payload)
-                );
-
-        SinkRecord sinkRecord = new SinkRecord(
-                "topic",
-                0,
-                SchemaBuilder.STRING_SCHEMA,
-                "key",
-                valueSchema,
-                value,
-                0
-        );
-
-        ConnectRecord transformedRecord = maskJsonField.apply(sinkRecord);
-
-        Struct data = (Struct)transformedRecord.value();
-        String replacedJson = data.getStruct("inner").getString("document");
-        //assertValue(replacedJson, path, replacement);
-
-        Assertions.assertEquals(expectedJson, replacedJson);
-    }
     @Test
     public void jsonPointerReplaceIntTest() throws IOException {
         replaceWithPointerLib(
@@ -640,5 +192,30 @@ public class MaskJsonFieldTest {
         }
 
         return root;
+    }
+
+    @Test
+    public void mapTest() {
+        Map<String,Object> map = new HashMap<>();
+        map.put("string", "string");
+        map.put("int", 1);
+        String [] array_string = new String[2];
+        array_string[0] = "first_string";
+        array_string[1] = "secibd_string";
+        map.put("array_string", array_string);
+        Integer [] array_int = new Integer[2];
+        array_int[0] = 100;
+        array_int[1] = 200;
+        map.put("array_int", array_int);
+
+        try {
+            String output = mapper.writeValueAsString(map);
+            System.out.println(output);
+
+            JsonNode node = mapper.readTree(output);
+        } catch (JsonProcessingException e) {
+            throw new RuntimeException(e);
+        }
+
     }
 }
